requirements.txt
python-dotenv
youtube-transcript-api
langchain
langchain-community
langchain-huggingface
langchain-google-genai
sentence-transformers
faiss-cpu
transformers
tqdm
